+++ tests/test_utils_xml.py
@@ -4,25 +4,65 @@ from fhir_core import xml_utils
 from tests.fixtures.resources.patient import Patient
 
 from .fixtures import (
-    STATIC_PATH,
     STATIC_PATH_JSON_EXAMPLES,
     STATIC_PATH_XML_EXAMPLES,
     STATIC_PATH_XML_SCHEMA,
+    FhirPrimitiveTypesModel,
 )
 
 __author__ = "Md Nazrul Islam"
 __email__ = "email2nazrul@gmail.com>"
 
 
+def test_xml_represent():
+    """ """
+    model = FhirPrimitiveTypesModel(
+        booleanTypeRequired=False,
+        stringTypeRequired="My String",
+        stringListTypeRequired=["line"],
+        uuidTypeRequired="0f617818-ee6a-4d11-86ff-cf5924038f27",
+        dateTypeRequired="1970-01-01",
+        base64BinaryTypeRequired="QUxMLUlOLU9ORQ==",
+        urlTypeNotRequired="https://schemas.xmlsoap.org/soap/encoding/",
+    )
+    assert (
+        xml_utils.xml_represent(
+            model.model_fields["booleanTypeRequired"], model.booleanTypeRequired
+        )
+        == "false"
+    )
+    assert (
+        xml_utils.xml_represent(
+            model.model_fields["base64BinaryTypeRequired"],
+            model.base64BinaryTypeRequired,
+        ).strip()
+        == "QUxMLUlOLU9ORQ=="
+    )
+
+    assert (
+        xml_utils.xml_represent(
+            model.model_fields["uuidTypeRequired"], model.uuidTypeRequired
+        )
+        == "urn:uuid:0f617818-ee6a-4d11-86ff-cf5924038f27"
+    )
+    assert (
+        xml_utils.xml_represent(
+            model.model_fields["urlTypeNotRequired"], model.urlTypeNotRequired
+        )
+        == "https://schemas.xmlsoap.org/soap/encoding/"
+    )
+    assert (
+        xml_utils.xml_represent(
+            model.model_fields["dateTypeRequired"], model.dateTypeRequired
+        )
+        == "1970-01-01"
+    )
+
+
 def offtest_xml_node_patient_resource():
-    """Accept-Charset: utf-8
-    Accept: application/fhir+xml;q=1.0, application/xml+fhir;q=0.9
-    User-Agent: HAPI-FHIR/5.3.0 (FHIR Client; FHIR 4.0.1/R4; apache)
-    Accept-Encoding: gzip
-    Content-Type: application/fhir+xml; charset=UTF-8
-    """
+    """ """
     patient_fhir = Patient.model_validate_json(
-        (STATIC_PATH_JSON_EXAMPLES / "Patient-with-ext.json").read_bytes()
+        (STATIC_PATH_JSON_EXAMPLES / "patient-example.json").read_bytes()
     )
     patient_node = xml_utils.Node.from_fhir_obj(patient_fhir)
 
@@ -34,12 +74,14 @@ def offtest_xml_node_patient_resource():
         raise AssertionError("code should not come here!")
 
 
-def offtest_element_to_node():
+def test_element_to_node():
     """ """
     schema = lxml.etree.XMLSchema(file=str(STATIC_PATH_XML_SCHEMA / "patient.xsd"))
     xmlparser = lxml.etree.XMLParser(schema=schema)
     element = lxml.etree.fromstring(
-        (STATIC_PATH / "Patient-with-ext.xml").read_bytes(),
+        (
+            STATIC_PATH_XML_EXAMPLES / "patient-example-f001-pieter(f001).xml"
+        ).read_bytes(),
         parser=xmlparser,
     )
     patient_node = xml_utils.Node.from_element(element)
@@ -49,25 +91,36 @@ def offtest_element_to_node():
         raise AssertionError("Code should not come here!")
 
 
-def offtest_model_obj_xml_file():
+def test_model_obj_xml_file():
     """ """
     patient = xml_utils.xml_loads(
-        Patient, (STATIC_PATH_XML_EXAMPLES / "Patient-with-ext.xml").read_bytes()
+        Patient,
+        (
+            STATIC_PATH_XML_EXAMPLES
+            / "patient-example-sex-and-gender(patient-example-sex-and-gender).xml"
+        ).read_bytes(),
     )
     # with parser parameter
     schema = lxml.etree.XMLSchema(file=str(STATIC_PATH_XML_SCHEMA / "patient.xsd"))
     xmlparser = lxml.etree.XMLParser(schema=schema)
     patient2 = xml_utils.xml_loads(
         Patient,
-        (STATIC_PATH_XML_EXAMPLES / "Patient-with-ext.xml").read_bytes(),
+        (
+            STATIC_PATH_XML_EXAMPLES
+            / "patient-example-sex-and-gender(patient-example-sex-and-gender).xml"
+        ).read_bytes(),
         xmlparser,
     )
     patient3 = xml_utils.xml_loads(
-        Patient, (STATIC_PATH_XML_EXAMPLES / "Patient-with-ext.xml").read_bytes()
+        Patient,
+        (
+            STATIC_PATH_XML_EXAMPLES
+            / "patient-example-sex-and-gender(patient-example-sex-and-gender).xml"
+        ).read_bytes(),
     )
     assert patient == patient2
     patient3.text = None
     patient.text = None
-    patient.contained[1].text = None
-    patient3.contained[1].text = None
+    # patient.contained[0].text = None
+    # patient3.contained[0].text = None
     assert patient3 == patient